remove milestone
after a long think about the scrum process of our team, i decide to remove milestone tag.
to every project, the milestone is necessary, but here, when we work on scrum, the date is
milestone. we may drop task, but we will never delay the milestone. So a set of milestone
should be binded on date, like event, but not tag.